refactor. Broke models and routes into seperate directories. replaced app. with router. //
