+++ models/tags.js
@@ -1,0 +1,8 @@
+var mongoose = require("mongoose");
+
+const tagSchema = new mongoose.Schema({
+	title: String,
+});
+
+
+module.exports = mongoose.model("Tag", tagSchema);
