+++ models/users.js
@@ -1,0 +1,22 @@
+var mongoose = require("mongoose");
+var passportLocalMongoose = require("passport-local-mongoose");
+
+
+const UserSchema = new mongoose.Schema({
+	username: {
+		type: String,
+		unique: true,
+		required: true
+	},
+	email: {
+		type: String,
+		unique: true,
+		required: true
+	},
+	password: String
+});
+
+UserSchema.plugin(passportLocalMongoose);
+
+module.exports = mongoose.model("user", UserSchema);
+
